migrate VVC box to separate file
