+++ libheif/file.cc
@@ -23,6 +23,7 @@
 #include "libheif/heif.h"
 #include "libheif/jpeg2000.h"
 #include "libheif/jpeg.h"
+#include "libheif/vvc.h"
 
 #include <cstdint>
 #include <fstream>
